Update user_queue table amount field closes #117

+++ src/main/kotlin/com/innopolis/innoqueue/dto/UserExpensesDTO.kt
@@ -11,7 +11,7 @@ data class UserExpensesDTO(
     @JsonProperty("user")
     val userName: String,
     @JsonProperty("expenses")
-    val expenses: Double?,
+    val expenses: Long?,
     @JsonProperty("is_active")
     val isActive: Boolean
 )
